Add files via upload
1.1	Purpose Of Project
The aim of this project is to generate 50 random 4-digit numbers, save these numbers as sold tickets in the txt called tickets.txt into the SD Card, then draw a random number from these 50 numbers and print the winning number on the LCD screen and the time the ticket was drawn with RTC. Then save this data as a log back into the SD Card.
1.2	Used materials
•	1 x Arduino Nano
•	1 x 16x2 LCD Display with I2C module
•	1 x RTC1307 with battery 
•	1 x Micro SD Card Module
•	1 x Mini Breadboard
•	1 x Micro SD Card
•	3 x Push Button
•	3 x 1k Ohm Resistance
•	1 x Data Cable
•	20 x Female-Male Jumper Cable
•	20 x Male-Male Jumper Cable

1.3	Fritzing Schema

 
1.4	Modules and Connections
SD Card Module: It works with 4.5-5 V.
•	GND  Ground
•	VCC  5V
•	MISO  D12
•	MOSI  D11
•	SCK  D13
•	CS  D10
RTC1307: It works with 3V so I wired to 3.3V on Arduino.
•	GND  Ground
•	VCC  3.3V
•	SDA  A4 (Used for data communication, A4 Default SDA port on Arduino Nano)
•	SCL  A5 (SCL used for synchronization of sending and receiving data parties)
I2C Module: It works with 5V. It reduces number of pin that LCD Display usage. 
•	GND  Ground
•	VCC  5V
•	SDA  A4 (Used for data communication, A4 Default SDA port on Arduino Nano)
•	SCL A5(SCL used for synchronization of sending and receiving data parties)
Buttons: They work with low voltage so I wired to 3.3V of Arduino.
•	D5,D6,D7
1.5	Functions of Project
 
Figure 1
Button1: Calls mainmenu function in Arduino which contains Opening Screen as shown in Figure 1.
 
Figure 2
Button 2: Draws a lot and displays on LCD. The rightmost number indicates which ticket the ticket is. Also displays Date time on the screen. Figure 2.
 
Figure 3
Button 3: Saves Winner number and date time of the draw into SD Card and returns to mainmenu function. Figure 3.